[redshift-gtk] Code refactor
- Refactored the classes into separate files
